refactor(cmd): extract daemon mode logic to daemon.go
